rpc_sasl.py: decode auth mechanism retrieved from the sasl library
While trying snakebite-py3 with other Kerberos libraries, I noticed
that SASL initialize was not able to use one of the available auth
mechanisms announced by the HDFS Namenode. The issue seems related
to a comparison between a text string and a bytes string.

In order to properly test/use this new code a new Kerberos library
must be chosen, but it will be done in a later commit.

Issue: #5